refactor: update for comment
