+++ packages/fx-core/src/common/azure-hosting/botServiceHosting.ts
@@ -1,0 +1,9 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+import { AzureHosting } from "./azureHosting";
+
+export class BotHosting extends AzureHosting {
+  configurable = false;
+  hostType = "botservice";
+}
