+++ packages/fx-core/src/common/azure-hosting/interface.ts
@@ -1,0 +1,10 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+export enum HostType {
+  AppService = "webapp",
+  Function = "function",
+  BotService = "botservice",
+}
+
+export type BicepConfigs = string[];
