+++ packages/fx-core/src/common/azure-hosting/utils.ts
@@ -1,19 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { Inputs, ResourceTemplate } from "@microsoft/teamsfx-api";
-import { Context } from "@microsoft/teamsfx-api/build/v2";
-import * as fs from "fs-extra";
-import path from "path";
-import { BotHostTypes, generateBicepFromFile } from "../../../../common";
-import { ArmTemplateResult } from "../../../../common/armInterface";
-import { Bicep } from "../../../../common/constants";
-import { getTemplatesFolder } from "../../../../folder";
-import { getActivatedV2ResourcePlugins } from "../../../solution/fx-solution/ResourcePluginContainer";
-import { NamedArmResourcePluginAdaptor } from "../../../solution/fx-solution/v2/adaptor";
-import { HostTypes, PluginBot } from "../resources/strings";
-import { BicepConfigs } from "./botSolution";
-import * as utils from "../utils/common";
+import { ArmTemplateResult } from "../armInterface";
 
 export function mergeTemplates(templates: ArmTemplateResult[]): ArmTemplateResult {
   const result: ArmTemplateResult = {
