+++ packages/fx-core/templates/plugins/resource/hosting/bicep/botservice/botserviceProvision.template.bicep
