+++ packages/fx-core/templates/plugins/resource/hosting/bicep/botservice/parameters.json
