+++ packages/fx-core/templates/plugins/resource/hosting/bicep/botservice/provision.template.bicep
