+++ packages/fx-core/templates/plugins/resource/hosting/bicep/function/config.template.bicep
