+++ packages/fx-core/templates/plugins/resource/hosting/bicep/function/functionConfiguration.template.bicep
