+++ packages/fx-core/templates/plugins/resource/hosting/bicep/function/functionProvision.template.bicep
