+++ packages/fx-core/templates/plugins/resource/hosting/bicep/function/provision.template.bicep
